project4/requirements.txt
pycryptodome==3.20.0
pyodbc==5.1.0